requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
python-dotenv==1.0.0
scikit-learn==1.4.0
joblib==1.3.2
pandas==2.2.0
numpy==1.26.4
tldextract==5.1.1
vt-py==0.14.0
requests==2.28.2
httpx==0.26.0
Pillow==10.2.0